backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==12.0.0
django-cloudinary-storage
django-json-widget
